change type from return value from float to double to be consistent

+++ src/main/java/eu/ecoepi/iris/components/Precipitation.java
@@ -15,7 +15,7 @@ public class Precipitation extends Component {
         this.rainfall = rainfall;
     }
 
-    public double getRainfall() {
+    public float getRainfall() {
         return rainfall;
     }
